+++ TIFoundationUtils/AsyncOperation/Sources/AsyncOperation+Observe.swift
@@ -27,18 +27,25 @@ private final class ClosureObserverOperation<Output, Failure: Error>: AsyncOpera
 
     public init(dependency: AsyncOperation<Output, Failure>,
                 onSuccess: ((Output) -> Void)? = nil,
-                onFailure: ((Failure) -> Void)? = nil) {
+                onFailure: ((Failure) -> Void)? = nil,
+                callbackQueue: DispatchQueue = .main) {
 
         super.init()
 
         cancelOnCancellation(of: dependency)
 
-        dependencyObservation = dependency.subscribe { [weak self] in
-            onSuccess?($0)
-            self?.handle(result: $0)
-        } onFailure: { [weak self] in
-            onFailure?($0)
-            self?.handle(error: $0)
+        dependencyObservation = dependency.subscribe { [weak self] result in
+            callbackQueue.async {
+                onSuccess?(result)
+            }
+
+            self?.handle(result: result)
+        } onFailure: { [weak self] error in
+            callbackQueue.async {
+                onFailure?(error)
+            }
+
+            self?.handle(error: error)
         }
 
         addDependency(dependency) // keeps strong reference to dependency as well
@@ -49,8 +56,12 @@ private final class ClosureObserverOperation<Output, Failure: Error>: AsyncOpera
 
 public extension AsyncOperation {
     func observe(onSuccess: ((Output) -> Void)? = nil,
-                 onFailure: ((Failure) -> Void)? = nil) -> AsyncOperation<Output, Failure> {
+                 onFailure: ((Failure) -> Void)? = nil,
+                 callbackQueue: DispatchQueue = .main) -> AsyncOperation<Output, Failure> {
 
-        ClosureObserverOperation(dependency: self, onSuccess: onSuccess, onFailure: onFailure)
+        ClosureObserverOperation(dependency: self,
+                                 onSuccess: onSuccess,
+                                 onFailure: onFailure,
+                                 callbackQueue: callbackQueue)
     }
 }