+++ TIMoyaNetworking/Sources/RecoverableNetworkService/DefaultRecoverableJsonNetworkService.swift
@@ -28,7 +28,7 @@ import Moya
 open class DefaultRecoverableJsonNetworkService<ApiError: Decodable & Error>: DefaultJsonNetworkService {
     public typealias ErrorHandler = AnyAsyncEventHandler<ApiError>
 
-    private var defaultErrorHandlers: [ErrorHandler] = []
+    private(set) public var defaultErrorHandlers: [ErrorHandler] = []
 
     public func process<B: Encodable, S: Decodable>(request: EndpointRequest<B, S>,
                                                     prependErrorHandlers: [ErrorHandler] = [],