+++ TINetworking/Sources/Mapping/Nothing.swift
@@ -20,12 +20,6 @@
 //  THE SOFTWARE.
 //
 
-public struct Parameter<Location: ParameterLocation> {
-    public let value: Any?
-    public let allowEmptyValue: Bool
-
-    public init(value: Any?, allowEmptyValue: Bool = false) {
-        self.value = value
-        self.allowEmptyValue = allowEmptyValue
-    }
+public struct Nothing: Codable {
+    public init() {}
 }